+++ API/Controllers/BookingController.cs
@@ -61,14 +61,17 @@ namespace API.Controllers
         }
         // Menghapus Data Booking
         [HttpDelete] // http method
-        public IActionResult Delete(Booking booking) 
+        public IActionResult Delete(Guid guid)
         {
+     
+            var booking = _bookingRepository.GetByGuid(guid);
+            if (booking is null)
+                return NotFound("Id Not Found");
+
             var result = _bookingRepository.Delete(booking);
-            if (!result) // return result bool true jika berhasil maka memakai negasi untuk gagal
-            {
-                return BadRequest("Failed to delete data"); // 400 dengan pesan
-            }
-            return Ok(result); //200 berhasil
-        }
+            if (!result)
+                return BadRequest("Failed to delete data");
+            return Ok(result);
+        } 
     }
 }
