+++ API/Controllers/EducationController.cs
@@ -60,13 +60,14 @@ namespace API.Controllers
         }
         // Menghapus Data Education
         [HttpDelete] // http method
-        public IActionResult Delete(Education education) 
-        { 
+        public IActionResult Delete(Guid guid) 
+        {
+            var education = _educationRepository.GetByGuid(guid);
+            if (education is null)
+                return NotFound("Id Not Found");
             var result = _educationRepository.Delete(education);
             if (!result) // return result bool true jika berhasil maka memakai negasi untuk gagal
-            {
                 return BadRequest("Failed to delete data"); // 400 dengan pesan
-            }
             return Ok(result); //200 berhasil
         }
     }
