+++ API/Controllers/RoleController.cs
@@ -60,13 +60,15 @@ namespace API.Controllers
         }
         // Menghapus Data Role
         [HttpDelete] // http method
-        public IActionResult Delete(Role role)
+        public IActionResult Delete(Guid guid)
         {
+            var role = _roleRepository.GetByGuid(guid);
+            if (role is null)
+                return NotFound("Id Not Found");
             var result = _roleRepository.Delete(role);
             if (!result) // return result bool true jika berhasil maka memakai negasi untuk gagal
-            {
                 return BadRequest("Failed to delete data"); // 400 dengan pesan
-            }
+
             return Ok(result); //200 berhasil
         }
     }
